service: Add ChatOpenAttachment d-bus method
The way we did this previously was to use the ExternalEvent method,
though ExternalEvent is not really for that, especially when
we know which message we are responding to. The event names
are really an implementation detail in coding-game-service
and external apps shouldn't really have to care about them.

Now, we examine open_event and treat it as an array where we can
dispatch multiple events upon response.