controller: Refactor out getStateForMissionName into separate func
Previously we were calling _startMission unconditionally every time
the chatbox started up. The previous assumption was that calling
_startMission on an already-started mission was a no-opt, but that
was invalidated when we added the CHANGE actor since there was a need
to return to already-completed or started missions there. That was
causing duplicate messages to appear every time the game service
was killed and restarted (since it would always try and restart
the currently active mission).

https://phabricator.endlessm.com/T14746