controller: Don't try and get settings schema in controller
Move this logic to communicator, where it can be isolated out for
the tests. Some tests were failing locally because of an uninstalled
showmehow-data (failed to get schema). In this case we don't care.

Unfortunately, this means that we need to look up the schema
both when the value is fetched and when it is set.